catArticulos

+++ ProyectoResidencias/Catalogos/Cat.Articulos/CatArticulos.cs
@@ -46,30 +46,45 @@ namespace ProyectoResidencias.Catalogos.Cat.Articulos
 
         private void toolStripButton3_Click(object sender, EventArgs e)
         {
-            string ConnString = Clases.stconexion.scon;
-            string SqlString = "Delete from CatArticulos where Descripcion='"+ Clases.Variables.referencia+"'";
-            try
-            {
-                SqlConnection conn = new SqlConnection(ConnString);
-                SqlCommand cmd = new SqlCommand(SqlString, conn);
-                cmd.CommandType = CommandType.Text;
-                conn.Open();
-                cmd.ExecuteNonQuery();
-                conn.Close();
-                MessageBox.Show("Articulo eliminado correctamente.");
-                Clases.LLenadoGrids.llenarCatArticulos(GVCatArticulos, "vArticulosCompras");
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show("El valor seleccionado no es valido. \n" + ex.ToString());
+            FConfirmacion fConfirmacion = new FConfirmacion();
+            fConfirmacion.ShowDialog();
+            if (fConfirmacion.DialogResult==DialogResult.OK) {
+                string ConnString = Clases.stconexion.scon;
+                string SqlString = "Delete from CatArticulos where Descripcion='" + Clases.Variables.referencia + "'";
+                try
+                {
+                    SqlConnection conn = new SqlConnection(ConnString);
+                    SqlCommand cmd = new SqlCommand(SqlString, conn);
+                    cmd.CommandType = CommandType.Text;
+                    conn.Open();
+                    cmd.ExecuteNonQuery();
+                    conn.Close();
+                    MessageBox.Show("Articulo eliminado correctamente.");
+                    Clases.LLenadoGrids.llenarCatArticulos(GVCatArticulos, "vArticulosCompras");
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("El valor seleccionado no es valido. \n" + ex.ToString());
+                }
             }
         }
 
         private void GVCatArticulos_CellEnter(object sender, DataGridViewCellEventArgs e)
         {
             var filaSeleccionada = GVCatArticulos.CurrentRow;
-            Clases.Variables.referencia = filaSeleccionada.Cells[0].Value.ToString();
-            Clases.Variables.desc = filaSeleccionada.Cells[1].Value.ToString();
+            Clases.Variables.referencia = filaSeleccionada.Cells[1].Value.ToString();
+            Clases.Variables.desc = filaSeleccionada.Cells[0].Value.ToString();
+            Clases.Variables.descA = filaSeleccionada.Cells[5].Value.ToString();
+            Clases.Variables.descF = filaSeleccionada.Cells[2].Value.ToString();
+            Clases.Variables.descM = filaSeleccionada.Cells[4].Value.ToString();
+            Clases.Variables.CHArticuloContrato = filaSeleccionada.Cells[6].Value.ToString();
+            Clases.Variables.CHInventariable=filaSeleccionada.Cells[3].Value.ToString();
+        }
+        
+        private void GVCatArticulos_CellDoubleClick(object sender, DataGridViewCellEventArgs e)
+        {
+            Catalogos.Cat.Articulos.Botones.Modificar modificar = new Botones.Modificar();
+               modificar.ShowDialog();
         }
     }
 }
